+++ dejafu/Test/DejaFu.hs
@@ -242,17 +242,18 @@ module Test.DejaFu
   , gives'
   ) where
 
-import           Control.Arrow      (first)
-import           Control.DeepSeq    (NFData(..))
-import           Control.Monad      (unless, when)
-import           Control.Monad.Ref  (MonadRef)
-import           Control.Monad.ST   (runST)
-import           Data.Function      (on)
-import           Data.List          (intercalate, intersperse, minimumBy)
-import           Data.Ord           (comparing)
+import           Control.Arrow        (first)
+import           Control.DeepSeq      (NFData(..))
+import           Control.Monad        (unless, when)
+import           Control.Monad.Ref    (MonadRef)
+import           Control.Monad.ST     (runST)
+import           Data.Function        (on)
+import           Data.List            (intercalate, intersperse, minimumBy)
+import           Data.Ord             (comparing)
 
 import           Test.DejaFu.Common
 import           Test.DejaFu.Conc
+import           Test.DejaFu.Defaults
 import           Test.DejaFu.SCT
 
 
@@ -687,60 +688,6 @@ gives expected results = go expected [] results $ defaultFail failures where
 -- @since 0.2.0.0
 gives' :: (Eq a, Show a) => [a] -> Predicate a
 gives' = gives . map Right
-
-
--------------------------------------------------------------------------------
--- Defaults
-
--- | A default way to execute concurrent programs: systematically
--- using 'defaultBounds'.
---
--- @since 0.6.0.0
-defaultWay :: Way
-defaultWay = Systematically defaultBounds
-
--- | The default memory model: @TotalStoreOrder@
---
--- @since 0.2.0.0
-defaultMemType :: MemType
-defaultMemType = TotalStoreOrder
-
--- | All bounds enabled, using their default values.
---
--- @since 0.2.0.0
-defaultBounds :: Bounds
-defaultBounds = Bounds
-  { boundPreemp = Just defaultPreemptionBound
-  , boundFair   = Just defaultFairBound
-  , boundLength = Just defaultLengthBound
-  }
-
--- | A sensible default preemption bound: 2.
---
--- See /Concurrency Testing Using Schedule Bounding: an Empirical Study/,
--- P. Thomson, A. F. Donaldson, A. Betts for justification.
---
--- @since 0.2.0.0
-defaultPreemptionBound :: PreemptionBound
-defaultPreemptionBound = 2
-
--- | A sensible default fair bound: 5.
---
--- This comes from playing around myself, but there is probably a
--- better default.
---
--- @since 0.2.0.0
-defaultFairBound :: FairBound
-defaultFairBound = 5
-
--- | A sensible default length bound: 250.
---
--- Based on the assumption that anything which executes for much
--- longer (or even this long) will take ages to test.
---
--- @since 0.2.0.0
-defaultLengthBound :: LengthBound
-defaultLengthBound = 250
 
 
 -------------------------------------------------------------------------------